Refactor angle into units package
