+++ src/main/scala/libra/dimension/Angle.scala
@@ -1,4 +1,5 @@
 package libra
+package dimension
 
 import libra.ops.quantity.ConvertTo
 import ops.base.{Conversion, ConversionFactor, Show}
@@ -10,7 +11,7 @@ import singleton.ops._
 import libra.si.{Second, Time}
 import shapeless._
 
-trait AngleDimension {
+trait Angle {
 
   type Angle
 
@@ -115,41 +116,4 @@ trait AngleDimension {
   ): ConversionFactor[A, Angle, Turn, Arcsecond] =
     gradianArcsecondConversion.compose(turnGradianConversion)
 
-}
-
-trait AngleDegree {
-  implicit final class AngleDegreeOps[A](val a: A) {
-    def degree: QuantityOf[A, Angle, Degree] = Quantity(a)
-  }
-}
-
-trait AngleArcminute {
-  implicit final class AngleArcminuteOps[A](val a: A) {
-    def arcminute: QuantityOf[A, Angle, Arcminute] = Quantity(a)
-  }
-}
-
-trait AngleArcsecond {
-  implicit final class ArcsecondOps[A](val a: A) {
-    def arcsecond: QuantityOf[A, Angle, Arcsecond] = Quantity(a)
-  }
-}
-
-
-trait AngleRadian {
-  implicit final class RadianOps[A](val a: A) {
-    def radian: QuantityOf[A, Angle, Radian] = Quantity(a)
-  }
-}
-
-trait AngleGradian {
-  implicit final class GradianOps[A](val a: A) {
-    def gradian: QuantityOf[A, Angle, Gradian] = Quantity(a)
-  }
-}
-
-trait AngleTurn {
-  implicit final class TurnOps[A](val a: A) {
-    def turn: QuantityOf[A, Angle, Turn] = Quantity(a)
-  }
 }
